chores: jewel chart modulize
